+++ FlowCrypt/Common/Errors.swift
@@ -10,18 +10,3 @@ enum Errors: Error {
     case programmingError(String)
     case valueError(String)
 }
-
-enum FCError: Error {
-    case general
-    case authentication
-    case operation(Error)
-}
-
-extension FCError {
-    init(_ error: Error) {
-        if (error as NSError).code == Imap.Err.authentication.rawValue {
-            self = .authentication
-        }
-        self = .operation(error)
-    }
-}
