+++ FlowCrypt/Controllers/Inbox/Views/InboxTableViewCell.xib
@@ -19,9 +19,9 @@
                 <autoresizingMask key="autoresizingMask"/>
                 <subviews>
                     <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="" textAlignment="right" lineBreakMode="tailTruncation" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="pge-Yk-wzv">
-                        <rect key="frame" x="170" y="20" width="190" height="25"/>
+                        <rect key="frame" x="270" y="20" width="90" height="25"/>
                         <constraints>
-                            <constraint firstAttribute="width" constant="190" id="21k-ly-n4U"/>
+                            <constraint firstAttribute="width" constant="90" id="21k-ly-n4U"/>
                             <constraint firstAttribute="height" constant="25" id="ePa-xg-nMy"/>
                         </constraints>
                         <fontDescription key="fontDescription" type="system" pointSize="17"/>
@@ -38,7 +38,7 @@
                         <nil key="highlightedColor"/>
                     </label>
                     <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="" lineBreakMode="tailTruncation" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="ykv-SI-g73">
-                        <rect key="frame" x="15" y="20" width="145" height="25"/>
+                        <rect key="frame" x="15" y="20" width="245" height="25"/>
                         <constraints>
                             <constraint firstAttribute="height" constant="25" id="Cnn-xK-3Xf"/>
                         </constraints>
